requirements.txt
pandas
numpy

sqlalchemy
psycopg2-binary

pyarrow

langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers

fastapi
uvicorn[standard]

tqdm
python-dotenv
